Implement RegenMeter into StatusOrbs
